# Conflicts:
#	src/main/java/com/findex/demo/indexData/index/domain/entity/IndexData.java
#	src/main/java/com/findex/demo/indexInfo/domain/entity/IndexInfo.java
#	src/main/java/com/findex/demo/syncJobs/domain/entity/SyncJob.java

+++ src/main/java/com/findex/demo/indexInfo/domain/entity/IndexInfo.java
@@ -13,14 +13,14 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class IndexInfo {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
 
-    private String indexClassification;
-    private String indexName;
-    private int employedItemCount;
-    private Instant basePointInTime;
-    private double baseIndex;
-    private boolean favorite;
+  private String indexClassification;
+  private String indexName;
+  private int employedItemCount;
+  private Instant basePointInTime;
+  private double baseIndex;
+  private boolean favorite;
 }
